+++ packages/assignment/src/@lib/components/notificationSystem.tsx
@@ -1,0 +1,31 @@
+// NotificationSystem 컴포넌트
+
+import { renderLog } from "../../utils";
+import { useNotifi } from "../provider/NotificationProvider";
+
+export const NotificationSystem: React.FC = () => {
+  const { notifications, removeNotification } = useNotifi();
+  renderLog('NotificationSystem rendered');
+
+  
+  return (
+    <div className="fixed bottom-4 right-4 space-y-2">
+      {notifications.map(notification => (
+        <div key={notification.id} className={`p-4 rounded shadow-lg ${
+          notification.type === 'success' ? 'bg-green-500' :
+            notification.type === 'error' ? 'bg-red-500' :
+              notification.type === 'warning' ? 'bg-yellow-500' : 'bg-blue-500'
+        } text-white`}>
+          {notification.message}
+          <button
+            onClick={() => removeNotification(notification.id)}
+            className="ml-4 text-white hover:text-gray-200"
+          >
+            닫기
+          </button>
+        </div>
+      ))}
+    </div>
+  );
+};
+export default NotificationSystem;
